% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ggToolsAndSettings.r
\name{ggR}
\alias{ggR}
\title{Plot rasters in ggplot with greyscale}
\usage{
ggR(x, layer = 1, maxpixels = 5e+06)
}
\arguments{
\item{x}{raster}

\item{layer}{layername}

\item{maxpixels}{Integer. Maximal number of pixels to sample}
}
\description{
Plot rasters in ggplot with greyscale
}

