% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggToolsAndSettings.r
\docType{data}
\name{MAPTHEME}
\alias{MAPTHEME}
\title{Theme for map plotting}
\format{\preformatted{List of 11
 $ axis.title.y     : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.title.x     : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.text.x      :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : chr "grey50"
  ..$ size      : num 8
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.text.y      :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : chr "grey50"
  ..$ size      : num 8
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : num 90
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.ticks       :List of 4
  ..$ colour  : chr "grey50"
  ..$ size    : NULL
  ..$ linetype: NULL
  ..$ lineend : NULL
  ..- attr(*, "class")= chr [1:2] "element_line" "element"
 $ axis.ticks.length:Class 'unit'  atomic [1:1] 0.06
  .. ..- attr(*, "unit")= chr "cm"
  .. ..- attr(*, "valid.unit")= int 1
 $ strip.text.x     :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 8
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ strip.text.y     :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 8
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : num 270
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ plot.title       :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 9
  ..$ hjust     : num 0
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ legend.title     :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 9
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ legend.text      :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 8
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
MAPTHEME
}
\description{
Theme for map plotting
}
\keyword{datasets}

