% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggToolsAndSettings.r
\name{SCALE}
\alias{SCALE}
\title{Scalebar for ggplot}
\usage{
SCALE(ras = NULL, len = 10000, yoff = 0.6, xoff = 0.2, sdist = 800,
  tdist = 1000, size = 1, col = "white")
}
\arguments{
\item{ras}{Raster*}

\item{len}{Numeric. Length of scalebar (map units)}

\item{yoff}{y offset from lower left corner (in percent)}

\item{xoff}{x offset from lower left corner (in percent)}

\item{sdist}{Numeric. Length of ticks}

\item{tdist}{Numeric. Distance text from ticks.}

\item{size}{Numeric. Thickness of scalebar}

\item{col}{Color of scalebar and text}
}
\description{
Scalebar for ggplot
}

