% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{knit2docx}
\alias{knit2docx}
\alias{knit2docx-package}
\title{knit2docx}
\usage{
knit2docx(.fileBasename, .docxFile = NULL, .withBibliography = TRUE,
  .bibFile = NULL, .bibStyle = "harvard1_mod.csl")
}
\arguments{
\item{.fileBasename}{character. basename of .Rmd file without extension, e.g. "Chapter1". Should be preceded by a path if not in the current working directory, e.g. "FolderX/Chapter1".}

\item{.docxFile}{optional. character. basename of output docx file. If not suplied \code{.fileBasename} will be used.}

\item{.withBibliography}{logical. run pandoc with or without pandoc-citeproc, i.e. with or without resolving citatations to a bibliography}

\item{.bibFile}{character. basename of Bibtex .bib file. Optional if equal to \code{x}.}

\item{.bibStyle}{literature formatting style}

\item{stopOnErrors}{logical. Stop if there were errors in R code. (Happens after knitr)}
}
\description{
knit2docx

Converts Rmd to docx including some sideeffects, such as figure labelling, file renaming etc. Currently settings
 are rather tightly tied to the requirements of pelagic publishing. Also it is not tested on anything else than linux.
 Pandoc and pandoc-citeproc must be installed and in your path. It is not overly beautiful, could be more efficient and less of a hack, but does the job.
}
\examples{
\dontrun{
## Copy demo files to knit2docx_demo folder
dir.create("knit2docx_demo")
file.copy(list.files(system.file("extdata/demo/", package = "knit2docx")), "knit2docx_demo", recursive = TRUE)
rlist.files("knit2docx_demo")
## Convert Rmd to docx
knit2docx("knit2docx_demo/Chapter_template")
## Open with Word:
list.files("knit2docx_demo", pattern = "docx")
## Remove demo directory
unlink("knit2docx_demo", recursive = TRUE)
}
}

